feat: 자료형 변경

+++ tingting/src/main/java/com/alsif/tingting/concert/dto/ConcertBaseDto.java
@@ -22,7 +22,7 @@ import lombok.ToString;
 public class ConcertBaseDto {
 
 	@Schema(description = "콘서트 PK", example = "1")
-	private Long concertSeq;
+	private Integer concertSeq;
 
 	@Schema(description = "콘서트 이름", example = "임영웅 2023 겨울 콘서트 - 서울")
 	private String name;
@@ -42,7 +42,7 @@ public class ConcertBaseDto {
 	@Schema(description = "콘서트홀 시도", example = "서울")
 	private String concertHallCity;
 
-	public ConcertBaseDto(Long concertSeq, String name, LocalDateTime holdOpenDate, LocalDateTime holdCloseDate,
+	public ConcertBaseDto(Integer concertSeq, String name, LocalDateTime holdOpenDate, LocalDateTime holdCloseDate,
 		String imageUrl,
 		String concertHallName, String concertHallCity) {
 		this.concertSeq = concertSeq;
